feat: Start adding state management via Vuex
There is still a lot to be done but the basic idea is that all character
information is represented in the Vuex state.
For now I don't differentiate between mutations and actions because all
changes are synchronous anyway.

Most of the modules will only have one corresponding box. For others,
some experimentation will be required whether it makes more sense to
use dynamic module registration or pass the index into an array of data
as a property of the box.