WIP: tentative printing support
This is all rather experimental. Firefox on Linux creates a decent-
looking PDF but it's not clear how it looks anywhere else.

Also, some concepts like the skills-box don't transfer to paper at all.

Supporting print support would probably be a major undertaking so I'll
have to re-think it.

In terms of the Agile Manifesto this is a huge opportunity to maximize
the work not done 😅